Replaced disabled symbols with explicit list
Previously, (burrow spectrum) and (burrow canvas) tried to define
a set of symbols no longer supplied by burrow owl, but supplied by
previous versions and could thus show up in older scripts found in
the wild.  This version tried to define these to give warnings, in
an effort to warn users that they were using an out-of-date script.
However, some of the symbols in the 'disabled' list were not in fact
disabled.  There was supposed to be a mechanism to detect which
symbols were actually not defined by burrow-owl, and only define
those as warnings.  This mechanism seems to be failing with
guile-2.0, so the list is replaced by an explicit list of symbols
which are actually not defined by the current version.